refactor: extract useSessions + useSplitLayout hooks from App.tsx
App.tsx had grown to 363 LOC owning four loosely-related concerns:
session state + IPC subscription wiring, terminal-instance refs, the
resizable split divider drag, and the layout composition. The two
non-presentational concerns are now their own hooks; App.tsx is back
to composition + the refit-on-active-switch effects.

App.tsx: 363 -> 168 LOC (-54%).

New hooks:

- src/useSessions.ts (208 LOC): owns sessions / statuses / activeId
  state, subscribes to onData / onShellData / onExit / onStatusChanged
  / onShellExit / onSessionAdded, catches up via listSessions and
  signals appReady. Exports closeSession / renameSession / newSession
  callbacks. Terminal refs are passed in (the caller still owns them
  so TerminalView and BottomTerminal can also receive them) and used
  by the dispose-on-remove path.

- src/useSplitLayout.ts (93 LOC): owns bottomHeight + persisted
  storage + the divider drag handler. Returns mainContainerRef so
  App.tsx can attach it to the split container.

Behaviour preserved verbatim — this is a mechanical move. The refit
effects (window resize, active-session switch) stay in App.tsx
because they touch both the active id from useSessions and the
terminal refs App.tsx still owns; co-locating them with composition
keeps the dependency clear.

No new tests — the refactor is behaviorally identical and the hooks
are tightly coupled to browser globals (window event listeners,
localStorage, document.body styles) that would require a jsdom test
harness to exercise. Manual smoke covers the integration; unit tests
for the pure helpers (clampHeight) already exist in src/layout.test.ts.

Verified: typecheck (both tsconfigs), npm test (98/98), npm run build.
